requirements.txt
requests
json
os
python-dotenv